third_party/blink/web_tests/external/wpt/css/css-pseudo/first-line-allowed-properties-expected.txt

This is a testharness.js-based test.
[FAIL] textJustify should be applied to first-line pseudo elements.
  assert_equals: expected "none" but got "auto"
Harness: the test ran to completion.
